refactor(CAG): * moved out & aliased getCagOutlinePaths * renamed isCagValid.js to cagValidation.js since it now contains more than one validator * more cleanups to CAG
