+++ src/core/utils/cagValidation.js
@@ -1,4 +1,5 @@
 const {areaEPS} = require('../constants')
+const {linesIntersect} = require('../math/lineUtils')
 
 // check if we are a valid CAG (for debugging)
 // NOTE(bebbi) uneven side count doesn't work because rounding with EPS isn't taken into account
@@ -36,4 +37,19 @@ const isCAGValid = function (CAG) {
   }
 }
 
-module.exports = isCAGValid
+const isSelfIntersecting = function (cag, debug) {
+  let numsides = cag.sides.length
+  for (let i = 0; i < numsides; i++) {
+    let side0 = cag.sides[i]
+    for (let ii = i + 1; ii < numsides; ii++) {
+      let side1 = cag.sides[ii]
+      if (linesIntersect(side0.vertex0.pos, side0.vertex1.pos, side1.vertex0.pos, side1.vertex1.pos)) {
+        if (debug) { console.log('side ' + i + ': ' + side0); console.log('side ' + ii + ': ' + side1) }
+        return true
+      }
+    }
+  }
+  return false
+}
+
+module.exports = {isCAGValid, isSelfIntersecting}
